LPS-50491 - Move block files in portalweb/portal/util/block to portalweb/util/block
